`StartEndRunButton`をToggleButtonベースのものにして、XAMLで書き直した
