+++ TRViS/DTAC/StartEndRunButton.xaml
@@ -1,0 +1,69 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ctrls:ToggleButton
+	xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+	xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+	x:Name="this"
+	xmlns:ctrls="clr-namespace:TRViS.Controls"
+	x:Class="TRViS.DTAC.StartEndRunButton">
+	<Frame
+		x:Name="BaseFrame"
+		CornerRadius="6"
+		Margin="0"
+		Padding="4"
+		VerticalOptions="Center"
+		HorizontalOptions="Center">
+		<HorizontalStackLayout
+			ScaleY="0.95"
+			HorizontalOptions="Center"
+			VerticalOptions="Center">
+			<Label
+				Text="&#xe039;"
+				TextColor="White"
+				FontFamily="MaterialIconsRegular"
+				FontSize="32"
+				VerticalOptions="Center"
+				Margin="2"
+				Padding="0">
+				<Label.Triggers>
+					<DataTrigger
+						TargetType="Label"
+						Binding="{Binding IsChecked, Source={x:Reference this}}"
+						Value="True">
+						<Setter
+							Property="Text"
+							Value="&#xe14b;"/>
+					</DataTrigger>
+				</Label.Triggers>
+			</Label>
+
+			<Label
+				Text="運行開始"
+				TextColor="White"
+				FontFamily="Hiragino Sans"
+				FontSize="24"
+				FontAttributes="Bold"
+				VerticalOptions="Center"
+				Margin="4"
+				Padding="0">
+				<Label.Triggers>
+					<DataTrigger
+						TargetType="Label"
+						Binding="{Binding IsChecked, Source={x:Reference this}}"
+						Value="True">
+						<Setter
+							Property="Text"
+							Value="運行終了"/>
+					</DataTrigger>
+				</Label.Triggers>
+			</Label>
+		</HorizontalStackLayout>
+	</Frame>
+
+	<ctrls:ToggleButton.Shadow>
+		<Shadow
+			Brush="Black"
+			Offset="3,3"
+			Radius="3"
+			Opacity="0.2"/>
+	</ctrls:ToggleButton.Shadow>
+</ctrls:ToggleButton>
