+++ TRViS/DTAC/VerticalStylePage.xaml
@@ -47,7 +47,9 @@
 			<dtac:StartEndRunButton
 				x:Name="StartEndRunButton"
 				Grid.Column="1"
-				Margin="2,8"/>
+				VerticalOptions="Center"
+				HorizontalOptions="End"
+				Margin="2"/>
 
 		</Grid>
